test: Move settings tests and improve didViewer test
