+++ src/screens/settings/__tests__/Settings.test.tsx
@@ -1,6 +1,6 @@
 import 'react-native'
 import React from 'react'
-import Settings from '../settings/Settings'
+import Settings from '../../settings/Settings'
 import { render, fireEvent } from 'react-native-testing-library'
 import { useTranslation } from 'react-i18next' //'../../../__mocks__/react-i18next'
